+++ Source/AppLovinAdapter.swift
@@ -126,7 +126,12 @@ final class AppLovinAdapter: PartnerAdapter {
         case .rewarded:
             return AppLovinAdapterRewardedAd(sdk: sdk, adapter: self, request: request, delegate: delegate)
         default:
-            throw error(.loadFailureUnsupportedAdFormat)
+            // Not using the `.adaptiveBanner` case directly to maintain backward compatibility with Chartboost Mediation 4.0
+            if request.format.rawValue == "adaptive_banner" {
+                return AppLovinAdapterBannerAd(sdk: sdk, adapter: self, request: request, delegate: delegate)
+            } else {
+                throw error(.loadFailureUnsupportedAdFormat)
+            }
         }
     }