+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Note the first digit of every adapter version corresponds to the major version o
 Adapters are compatible with any Chartboost Mediation SDK version within that major version.
 
 ### 4.11.11.0.1
+- Add support for Adaptive Banners.
+- No longer performing console logging on iOS 11.
 - This version of the adapters has been certified with AppLovinSDK 11.11.0.
 
 ### 4.11.11.0.0